Assign variable to empty string and loop 3 times
Make a new variable called outcome, and assign it an empty string
Then create for loop to increment num by 1 whilst it's less than three
Finally add outcome variable to string argument on each iteration of
the loop. Then outside the loop return the output. What this will do
is return String 3 times 'hey' resulting in the outcome of 'heyheyhey'