Complete the function to reverse the string
Add a new string variable called newWord with an empty value.
Create a for loop to loop through the entered string - 1.
Use word.length - 1 for this. We do this because the first char in the
string is at position 0, so we want to make sure we are starting
one letter back to ensure we are beginning at the last letter of
the string. As we loop through we are decrementing i by 1 each time,
so it moves back one letter.
Finally after each iteration of the loop we are adding our new
variable to the original string at position i, thus reversing each
letters original position through each iteration of the loop