Satisfy the final 3 test conditions
Improve the logic for checking whether num1, num2 is greater than
or less to by adding the Math.max and Math.min object. In order to
make this work we will need to store num1 under second variable "temp"
this will prevent the original value of num1 being
changed during the min.max section. Finally add a condition that checks
if the number entered is positive and to only accept an integer.