Shorten code by reducing two else if into a single else if
Combine the else if condition to check if leapYear1 is true, and
for notLeapYear to be false. Return true if both these conditions
match. We must make sure the condition for leapYear2 is checked first
because this is the definite rule, if the year is divisible by 400,
it is true every time, which is why this condition must be checked
first.
Finally add a else statement to return false if neither of
the if or else if conditions are true. Doing this will satisfy
all the test conditions.